+++ src/main/java/DTO/ReleaseRequest.java
@@ -1,0 +1,24 @@
+package DTO;
+
+import java.util.Date;
+
+public class ReleaseRequest {
+    private Date start;
+    private Date end;
+
+    public void setStart(Date start) {
+        this.start = start;
+    }
+
+    public void setEnd(Date end) {
+        this.end = end;
+    }
+
+    public Date getStart() {
+        return start;
+    }
+
+    public Date getEnd() {
+        return end;
+    }
+}
