+++ src/main/java/com/example/board/exception/ControllerAdvice.java
@@ -1,0 +1,16 @@
+package com.example.board.exception;
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+@Slf4j
+@org.springframework.web.bind.annotation.ControllerAdvice
+public class ControllerAdvice {
+    @ExceptionHandler(BusinessException.class)
+    public ResponseEntity<Response> handleException(BusinessException e) {
+        Response response = new Response(e.getMessage());
+        log.error("common error: {}" , e.toString());
+        return new ResponseEntity<>(response, HttpStatus.OK);
+    }
+}
