+++ src/main/java/com/example/board/exception/Response.java
@@ -1,0 +1,20 @@
+package com.example.board.exception;
+
+public class Response {
+    private String message;
+
+    public Response() {
+    }
+
+    public Response(String message) {
+        this.message = message;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+
+    public void setMessage(String message) {
+        this.message = message;
+    }
+}
